Fixed requiredAimTime and throwThreshold to be floats.

+++ Assets/Shared/Scripts/Network/ClinicianClasses/ClinicianClasses.cs
@@ -33,10 +33,10 @@ namespace SocketClasses
         public bool griplessGrabbing;
         public bool leftSideSpawnOnly;
         public string releaseButton;
-        public double requiredAimTime;
+        public float requiredAimTime;
         public bool rightSideSpawnOnly;
         public int targets;
-        public int throwThreshold;
+        public float throwThreshold;
         public bool useAutoAim;
         public bool useAutoReleaseTimerThrow;
         public bool useButtonPressForThrow;
@@ -82,10 +82,10 @@ namespace SocketClasses
         public static bool griplessGrabbing = false;
         public static bool leftSideSpawnOnly = false;
         public static string releaseButton = "Trigger";
-        public static double requiredAimTime = 3.0;
+        public static float requiredAimTime = 3.0;
         public static bool rightSideSpawnOnly = false;
         public static int targets = 1;
-        public static int throwThreshold = 35;
+        public static float throwThreshold = 35;
         public static bool useAutoAim = false;
         public static bool useAutoReleaseTimerThrow = false;
         public static bool useButtonPressForThrow = true;
